+++ Items/Shard.cpp
@@ -72,7 +72,7 @@ void Shard::InitShard(double r, double worldSizeIn) {
       double tmpangle = (M_PI * 2 / (double)npts * (double)j);
       double tmpRad = r * sin(tmpangle);
       double tmpH = r * cos(tmpangle) * 2.0;
-      int tmpPts = npts * (tmpRad / r);
+      int tmpPts = (int) (npts * (tmpRad / r));
       if (tmpPts < 0) {
          tmpPts *= -1;
       } else if (tmpPts == 0) {
